# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Supply Chain AI Platform API
  description: |
    Production-grade multi-agent AI supply chain orchestration API powered by Claude Opus 4.6.

    ## Features
    - **Multi-Agent Orchestration**: 7 specialized AI agents coordinated by Claude Opus 4.6
    - **Real-time Inventory Management**: Stock tracking with automated reorder recommendations
    - **Order Processing**: End-to-end order lifecycle with fraud detection
    - **Supplier Management**: Vendor evaluation and RFQ generation
    - **Demand Forecasting**: ML-based predictions using XGBoost
    - **Anomaly Detection**: Real-time fraud and anomaly scoring

    ## Authentication
    All endpoints require JWT Bearer token authentication (except `/health`, `/ready`, and `/api/v1/auth/token`).

    ```
    Authorization: Bearer <your-jwt-token>
    ```

    Obtain tokens via `POST /api/v1/auth/token` (OAuth2 password flow).

    ## Rate Limiting
    - 100 requests/minute per authenticated user
    - 1000 requests/hour per authenticated user
    - Agent orchestration: 10 requests/minute per user
  version: "1.0.0"
  contact:
    name: Platform Team
    email: platform@supply-chain.internal
  license:
    name: Proprietary
    url: https://supply-chain.internal/license
servers:
  - url: http://localhost:8000
    description: Local Development
  - url: https://api.supply-chain.internal
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/orders:
    get:
      tags: [Orders]
      summary: List orders
      operationId: list_orders
      parameters:
        - name: page
          in: query
          schema: {type: integer, default: 1, minimum: 1}
        - name: page_size
          in: query
          schema: {type: integer, default: 20, minimum: 1, maximum: 100}
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, processing, allocated, picking, packing, shipped, delivered, cancelled, returned, refunded]
        - name: search
          in: query
          schema: {type: string}
          description: Search by customer name, email, or order number
      responses:
        "200":
          description: Paginated order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"

    post:
      tags: [Orders]
      summary: Create order
      operationId: create_order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreate"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/orders/{order_id}:
    get:
      tags: [Orders]
      summary: Get order by ID
      operationId: get_order
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Orders]
      summary: Update order
      operationId: update_order
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Order updated

    delete:
      tags: [Orders]
      summary: Cancel order
      operationId: cancel_order
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Order cancelled
        "409":
          description: Order cannot be cancelled in current status
  /api/v1/orders/{order_id}/process:
    post:
      tags: [Orders]
      summary: Process order through AI agents
      operationId: process_order
      description: |
        Triggers multi-agent orchestration for order processing.
        Returns 202 Accepted immediately; use the task_id to poll for results.
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "202":
          description: Order processing initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id: {type: string}
                  status: {type: string, example: processing}
                  poll_url: {type: string}
  /api/v1/orders/bulk:
    post:
      tags: [Orders]
      summary: Create multiple orders
      operationId: bulk_create_orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orders:
                  type: array
                  items:
                    $ref: "#/components/schemas/OrderCreate"
                  maxItems: 100
      responses:
        "207":
          description: Multi-status response
...
  /api/v1/shipments/track/{tracking_number}:
    get:
      tags: [Shipments]
      summary: Track shipment
      operationId: track_shipment
      parameters:
        - name: tracking_number
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Shipment tracking info
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrderId:
      name: order_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc: {type: array, items: {type: string}}
              msg: {type: string}
              type: {type: string}
...
    OrderItem:
      type: object
      required: [sku, quantity, unit_price]
      properties:
        sku:
          type: string
        name:
          type: string
        quantity:
          type: integer
          minimum: 1
        unit_price:
          type: number
          minimum: 0
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        postal_code:
          type: string
        country:
          type: string
          default: "US"
    OrderCreate:
      type: object
      required: [customer_id, items]
      properties:
        customer_id:
          type: string
          format: uuid
        customer_email:
          type: string
          format: email
        customer_name:
          type: string
        priority:
          type: string
          enum: [low, standard, high, urgent]
          default: standard
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
          minItems: 1
        shipping_address:
          $ref: "#/components/schemas/Address"
        billing_address:
          $ref: "#/components/schemas/Address"
        notes:
          type: string
        requested_delivery_date:
          type: string
          format: date-time
        currency:
          type: string
          default: "USD"
    OrderUpdate:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string
        priority:
          type: string
    OrderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_number:
          type: string
        customer_id:
          type: string
          format: uuid
        customer_email:
          type: string
        customer_name:
          type: string
        status:
          type: string
        priority:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        tax_amount:
          type: number
        shipping_cost:
          type: number
        total_amount:
          type: number
        currency:
          type: string
        fraud_score:
          type: number
          description: "0.0-1.0, where >0.7 indicates high fraud risk"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderResponse"
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        pages:
          type: integer
...